prompt: Add a vte-urlencode utility
This allows us to add a tcsh counterpart of vte.sh, which is currently
missing and is hard to implement due to tcsh scripting limitations (no
support for functions).

An external utility is more expensive than a function, since it requires
a fork/exec, but __vte_urlencode in vte.sh already forks a subshell to
set LC_ALL to "C", which has almost the same cost.